+++ R/ggdotplotstats.R
@@ -123,7 +123,7 @@ ggdotplotstats <- function(data,
           test.value = test.value,
           bf.prior = bf.prior,
           top.text = caption,
-          output = "caption",
+          output = "expression",
           k = k
         )
     }
@@ -136,9 +136,7 @@ ggdotplotstats <- function(data,
         type = type,
         test.value = test.value,
         bf.prior = bf.prior,
-        robust.estimator = "onestep",
         effsize.type = effsize.type,
-        conf.type = "norm",
         conf.level = conf.level,
         nboot = nboot,
         k = k