+++ tests/testthat/test-ggwithinstats.R
@@ -591,3 +591,44 @@ testthat::test_that(
     testthat::expect_identical(p$labels$y, "Taste rating")
   }
 )
+
+# turning off mean path works ------------------------------------------
+
+testthat::test_that(
+  desc = "turning off mean path works",
+  code = {
+    set.seed(123)
+    library(ggstatsplot)
+
+    p1 <-
+      ggwithinstats(
+        iris_long,
+        condition,
+        value,
+        mean.point.args = list(size = 5, alpha = 0.5, color = "darkred"),
+        mean.path = TRUE,
+        results.subtitle = FALSE,
+        pairwise.comparisons = FALSE
+      )
+
+    p2 <-
+      ggwithinstats(
+        iris_long,
+        condition,
+        value,
+        mean.point.args = list(size = 5, alpha = 0.5, color = "darkred"),
+        mean.path = FALSE,
+        results.subtitle = FALSE,
+        pairwise.comparisons = FALSE
+      )
+
+    pb1 <- ggplot2::ggplot_build(p1)
+    pb2 <- ggplot2::ggplot_build(p2)
+
+    testthat::expect_equal(pb1$data[[1]], pb2$data[[1]])
+    testthat::expect_equal(pb1$data[[2]], pb2$data[[2]])
+    testthat::expect_equal(pb1$data[[3]], pb2$data[[3]])
+    testthat::expect_equal(pb1$data[[5]], pb2$data[[4]])
+    testthat::expect_equal(pb1$data[[6]], pb2$data[[5]])
+  }
+)